part1/main/api/book_project/requirements.txt
asgiref==3.10.0
Django==5.2
djangorestframework==3.16.1
Markdown==3.9
pillow==12.0.0
sqlparse==0.5.3
